Rename Directory to FileSystem
It looks like the extra flexibility to handle any subdirectory of
a file system as a file system in itself isn't really all that
useful. So I think it is simpler to just use a full file system
as the accessor and drop the support for returning objects that
access a subdirectory.