Use the subtree backup handlers from config directly in backupoperation
I decided copying all that data from the config object to the
backupoperation object was pretty silly, as the two representations
had to be pretty similar anyway. It increases coupling a bit (since
now both backupoperation and config depends on the same concrete class),
but it reduces duplicate and near-duplicate code quite a bit.

It also completely obsoletes the previous commit. Such is life.