Ignore symlinks
This is not optimal, as it discards the information that there was a
symlink there. It is probably ok not to back up the data pointed to
by the symlink. Any important data pointed to by a symlink should
(hopefully) have the pointed-to file in the source tree as well.