Avoid writing "small" files to temporary content storage before checking for duplicates
Though in practice it might be even better to never write any file to
content storage before checking for duplicates. That would mean that every
new file would have to be read twice, once for calculating the checksum and
once for writing to content storage. Every duplicate (assuming no checksum
collisions) would also have to be read twice, once to calculate the checksum
and once to compare to the duplicate. But I would save writing the file to
content storage and then deleting it again.