Fix the missing default logger service
The commit
   ce8e410 Make Logger print its logged messages by default
provided the service provider to the logger object. However, the logger
is one of the service provider's services, so the service provider setup
was split into the independent services and the services requiring other
services (currently only logger). However, the default services object
was returned before the second part was run, leaving the logger broken.