Make BackupData.get_file_info.mtime be consistent with [...].mtime_nsec
"Obviously" the right thing to do. And hopefully that will make the code
assume that files are unchanged when their mtime and size have not changed
since the previous backup. (I suspect that fails now because the mtime
from the stat includes the microsecond value while the one from the
database didn't.)

Unfortunately, the test_files_with_unchanged_mtime.... test doesn't
catch this, since it is a problem between units. I should probably
add a system test for that.