streamingreader: Correctly test for overrunning the block end
I actually fixed one of these when I put together a test that wrote
(and read) a block that exactly filled the available space. However,
I overlooked that there were more of these tests. There should be
tests for each of them, but it is just a little bit too much of a
pain to do that right now.